library: move plugin files to subdirectory
