+++ src/library/plugin/plugin_api.h
